Refactor configuration management: - Updated Router to use new configuration components. - Added AddConfiguration and EditConfiguration components for creating and editing configurations. - Implemented ViewConfiguration to display configuration details. - Adjusted API calls and state management for customization data. - Enhanced UI with improved loading states and error handling.
